extract temorary implementation of `ERR_INVALID_ARG_TYPE` class into a separate file
